[ALPS04262736] binder: fix race between munmap() and direct reclaim
An munmap() on a binder device causes binder_vma_close() to be called
which clears the alloc->vma pointer.

If direct reclaim causes binder_alloc_free_page() to be called, there
is a race where alloc->vma is read into a local vma pointer and then
used later after the mm->mmap_sem is acquired. This can result in
calling zap_page_range() with an invalid vma which manifests as a
use-after-free in zap_page_range().

The fix is to check alloc->vma after acquiring the mmap_sem (which we
were acquiring anyway) and bail out of binder_alloc_free_page() if it
has changed to NULL.

MTK-Commit-Id: 7aa7b50b074ff36aeebe0c0db8088579df7b25b8

Change-Id: I4d9b6554b47851ea0ddcf9f0adee446682f411e3
CR-Id: ALPS04262736
Feature: [Module]Binder
Signed-off-by: Frankie Chang <Frankie.Chang@mediatek.com>