[ALPS04718418] Kernel: Workaround for invalidating zero page
[Detail]
When user allocates a page without touching, this page would be
readonly, and if we invalidate this page in kernel without
write permission, then it will trigger permission fault and cause KE.

Workaround: instead of using "dc ivac", we use "dc civac" to be
a workaround. The solution should move the "user cache flush APIs" to
user space.

MTK-Commit-Id: 747df69a75d1a34d3702878f668bc3745fb2e599

Change-Id: I3bf3831413d65897e3ad8a53ec16363416b0f964
CR-Id: ALPS04718418
Feature: [Module]ION/M4U
Signed-off-by: Andy <chia-hao.hsu@mediatek.com>